airbnb/requirements.txt
requests
beautifulsoup4
openai
tweepy
schedule
pandas
matplotlib
praw
pytrends
PyYAML
streamlit
python-medium
python-dotenv
